requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.4
PyYAML==6.0.1
opencv-python==4.9.0.80
albumentations==1.4.3
scikit-learn==1.4.2
matplotlib==3.8.4
onnx==1.16.1
onnxruntime==1.17.3
click==8.1.7
python-dotenv==1.0.1
